attempt fix for directml
I have attempted to get directml running for my AMD gpu. It now uses the proper import and can detect the backend.

There is a problem however and seems to be just a limitation of what AMD gpus can do on windows. The gpu fails to be able to train the model because of the LSTM layer.
If somebody is able to get it working on AMD gpu in windows then please let me know so I can get it fixed.
For now I will be unable to train the model because of my current limitations.